% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{plot.scm_spillover}
\alias{plot.scm_spillover}
\title{Plot method for scm_spillover objects}
\usage{
\method{plot}{scm_spillover}(x, type = "effects", ...)
}
\description{
Plot method for scm_spillover objects
}
