% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmSpillover-package.R
\docType{package}
\name{scmSpillover-package}
\alias{scmSpillover}
\alias{scmSpillover-package}
\title{scmSpillover: Synthetic Control Method with Spillover Effects}
\description{
A general-purpose implementation of synthetic control methods that accounts for potential spillover effects between units. Based on the methodology of Cao and Dowd (2019).
}
\author{
\strong{Maintainer}: Zhanchao Fu \email{fuzhanchao2001@gmail.com}

Authors:
\itemize{
  \item Jianfei Cao
}

}
\keyword{internal}
